Move domains functions to a separate file
This commit only consists of renaming files and moving some functions.
